fix(mobile): address camera capture review findings
