+++ mobile/lib/features/profile/settings_profile_header.dart
@@ -8,6 +8,7 @@ import 'package:lucide_icons_flutter/lucide_icons.dart';
 import '../../shared/animated_avatar.dart';
 import '../../shared/custom_emoji/custom_emoji_provider.dart';
 import '../../shared/custom_emoji/custom_emoji_render.dart';
+import '../../shared/relay/media_image.dart';
 import '../../shared/theme/theme.dart';
 import '../../shared/widgets/avatar_image.dart';
 import '../../shared/widgets/anchored_popover_menu.dart';
@@ -34,6 +35,7 @@ class SettingsProfileHeader extends HookConsumerWidget {
     final hasStatus = status != null && !status.isEmpty;
     final presence = ref.watch(presenceProvider).value ?? 'offline';
     final animatedAvatar = parseAnimatedAvatarUrl(profile?.avatarUrl);
+    final animatedPosterUrl = animatedAvatar?.posterUrl;
     final handoff = ref.watch(profileAvatarHandoffProvider);
     final activeHandoff = handoff?.avatarUrl == profile?.avatarUrl
         ? handoff
@@ -84,15 +86,53 @@ class SettingsProfileHeader extends HookConsumerWidget {
                                   .read(profileAvatarHandoffProvider.notifier)
                                   .clear(activeHandoff.avatarUrl),
                       )
-                    : activeHandoff == null
+                    : activeHandoff == null || animatedPosterUrl == null
                     ? AvatarImageContent(
                         imageUrl: avatarUrl,
                         fallback: _AvatarFallback(initial: profile?.initial),
                       )
-                    : Image(
-                        image: MemoryImage(activeHandoff.poster),
-                        fit: BoxFit.cover,
-                        gaplessPlayback: true,
+                    : Stack(
+                        fit: StackFit.expand,
+                        children: [
+                          Image(
+                            image: MemoryImage(activeHandoff.poster),
+                            fit: BoxFit.cover,
+                            gaplessPlayback: true,
+                          ),
+                          Offstage(
+                            offstage: true,
+                            child: MediaImage(
+                              key: ValueKey(
+                                'settings-profile-paused-handoff-${activeHandoff.avatarUrl}',
+                              ),
+                              url: animatedPosterUrl,
+                              fit: BoxFit.cover,
+                              errorBuilder: (_, _, _) =>
+                                  const SizedBox.shrink(),
+                              frameBuilder:
+                                  (
+                                    context,
+                                    child,
+                                    frame,
+                                    wasSynchronouslyLoaded,
+                                  ) {
+                                    if (wasSynchronouslyLoaded ||
+                                        frame != null) {
+                                      WidgetsBinding.instance
+                                          .addPostFrameCallback((_) {
+                                            ref
+                                                .read(
+                                                  profileAvatarHandoffProvider
+                                                      .notifier,
+                                                )
+                                                .clear(activeHandoff.avatarUrl);
+                                          });
+                                    }
+                                    return child;
+                                  },
+                            ),
+                          ),
+                        ],
                       ),
               ),
             ),
